# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WKIL Backend API
  version: 1.0.0
  license:
    name: ISC
    identifier: ISC
  description: Contract for the Express v1 backend, including auth, Meta channels, content, media, widgets, admin, manager, analytics, integrations, and operational endpoints.
servers:
  - url: /
    description: Current API host
security:
  - bearerAuth: []
  - accessTokenCookie: []
...
paths:
...
  /v1/messenger/webhook:
    get:
      tags: [Messenger]
      operationId: verifyMessengerWebhook
      summary: Verify Messenger webhook subscription
      security: []
      parameters:
        - { name: hub.mode, in: query, required: false, schema: { type: string } }
        - { name: hub.verify_token, in: query, required: false, schema: { type: string } }
        - { name: hub.challenge, in: query, required: false, schema: { type: string } }
      responses:
        "200": { description: Webhook challenge or verification status. }
    post:
      tags: [Messenger]
      operationId: handleMessengerWebhook
      summary: Receive Messenger webhook events
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
  /v1/messenger:
    get:
      tags: [Messenger]
      operationId: listMessengerConversations
      summary: List Messenger conversations
      parameters:
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/CursorQuery"
        - $ref: "#/components/parameters/ChannelQuery"
        - $ref: "#/components/parameters/StatusQuery"
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/whatsapp/webhook:
    get:
      tags: [WhatsApp]
      operationId: verifyWhatsAppWebhook
      summary: Verify WhatsApp webhook subscription
      security: []
      parameters:
        - { name: hub.mode, in: query, required: false, schema: { type: string } }
        - { name: hub.verify_token, in: query, required: false, schema: { type: string } }
        - { name: hub.challenge, in: query, required: false, schema: { type: string } }
      responses:
        "200": { description: Webhook challenge or verification status. }
    post:
      tags: [WhatsApp]
      operationId: handleWhatsAppWebhook
      summary: Receive WhatsApp webhook events
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
...
  /v1/order-integrations/{integrationKey}/events:
    post:
      tags: [Order Confirmations]
      operationId: receiveOrderIntegrationEvent
      summary: Accept a signed canonical order event
      description: |
        Accepts raw JSON from an order source. The `Idempotency-Key` header must
        equal the top-level `eventId`. The HMAC is calculated over
        `${X-Wkil-Timestamp}.${rawRequestBody}` using the current signing secret;
        the previous signing secret is also accepted during secret rotation.
        The event is persisted before the `202` response and duplicate events
        are acknowledged without enqueueing another processing job.
      security: []
      parameters:
        - $ref: "#/components/parameters/IntegrationKeyPath"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
        - $ref: "#/components/parameters/WkilTimestampHeader"
        - $ref: "#/components/parameters/WkilSignatureHeader"
      requestBody:
        required: true
        description: Raw JSON bytes. Whitespace and key order are part of the signed payload.
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CanonicalOrderEvent" }
      responses:
        "202": { $ref: "#/components/responses/OrderEventAccepted" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }
  /v1/order-integrations:
    get:
      tags: [Order Confirmations]
      operationId: listOrderIntegrations
      summary: List order integrations
      description: Lists integrations for business profiles accessible to the authenticated user.
      security:
        - bearerAuth: []
        - accessTokenCookie: []
      parameters:
        - name: businessProfileId
          in: query
          required: false
          schema: { type: integer, minimum: 1 }
      responses:
        "200": { $ref: "#/components/responses/OrderIntegrationList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Order Confirmations]
      operationId: createOrderIntegration
      summary: Create an order integration
      description: |
        Creates a generic order integration. The signing secret is generated by
        the server and returned only once. If a callback secret is generated by
        rotation, it is also returned only once. Enabling store synchronization
        requires an HTTPS callback URL and a callback secret.
      security:
        - bearerAuth: []
          csrfHeader: []
        - accessTokenCookie: []
          csrfHeader: []
      requestBody: { $ref: "#/components/requestBodies/OrderIntegrationCreate" }
      responses:
        "201": { $ref: "#/components/responses/OrderIntegrationCreate" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/ServerError" }
...
  /v1/workspace:
    get:
      tags: [Business Profile]
      operationId: listWorkspaces
      summary: List workspaces available to the current user
      security:
        - bearerAuth: []
        - accessTokenCookie: []
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Business Profile]
      operationId: createWorkspace
      summary: Create a workspace for the current user
      security:
        - bearerAuth: []
          csrfHeader: []
        - accessTokenCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
      responses:
        "201": { $ref: "#/components/responses/Created" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/workspace/invite/{token}/info:
    get:
      tags: [Business Profile]
      operationId: getWorkspaceInviteInfo
      summary: Read public workspace invite information
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/facebook/analytics:
    get:
      tags: [Facebook]
      operationId: getFacebookAnalytics
      summary: Get Facebook analytics
      parameters:
        - $ref: "#/components/parameters/DaysQuery"
      responses:
        "200": { $ref: "#/components/responses/Ok" }
...
  /v1/leads:
    get:
      tags: [Leads]
      operationId: listLeads
      summary: List leads
      responses:
        "200": { $ref: "#/components/responses/Ok" }
    post:
      tags: [Leads]
      operationId: createLead
      summary: Create a lead
      security: []
      requestBody: { $ref: "#/components/requestBodies/Lead" }
      responses:
        "201": { $ref: "#/components/responses/Created" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/customers:
    get:
      tags: [Customers]
      operationId: listCustomers
      summary: List customers
      parameters:
        - { name: businessProfileId, in: query, required: false, schema: { type: integer } }
        - { name: q, in: query, required: false, schema: { type: string } }
        - { name: hasPhone, in: query, required: false, schema: { type: boolean } }
        - { name: status, in: query, required: false, schema: { type: string } }
        - { name: channel, in: query, required: false, schema: { type: string } }
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200": { $ref: "#/components/responses/Ok" }
...
  /v1/admin/users:
    get:
      tags: [Admin]
      operationId: adminListUsers
      summary: List users
      parameters:
        - { name: includeInactive, in: query, required: false, schema: { type: boolean } }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
...
  /v1/business-profile:
    get:
      tags: [Business Profile]
      operationId: listBusinessProfiles
      summary: List current user's business profiles
      responses:
        "200": { $ref: "#/components/responses/Ok" }
    post:
      tags: [Business Profile]
      operationId: createBusinessProfile
      summary: Create business profile
      security: [ { bearerAuth: [], csrfHeader: [] }, { accessTokenCookie: [], csrfHeader: [] } ]
      requestBody: { $ref: "#/components/requestBodies/BusinessProfile" }
      responses:
        "201": { $ref: "#/components/responses/Created" }
...
  /v1/media:
    get:
      tags: [Media]
      operationId: listMediaAssets
      summary: List media library assets
      parameters:
        - { name: businessProfileId, in: query, required: true, schema: { type: integer } }
        - { name: usageScope, in: query, required: false, schema: { $ref: "#/components/schemas/MediaUsageScope" } }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
    post:
      tags: [Media]
      operationId: uploadMediaAsset
      summary: Upload media library asset
      security: [ { bearerAuth: [], csrfHeader: [] }, { accessTokenCookie: [], csrfHeader: [] } ]
      requestBody: { $ref: "#/components/requestBodies/MediaUpload" }
      responses:
        "201": { $ref: "#/components/responses/Created" }
...
components:
  securitySchemes:
    mobileRefreshBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A WKIL refresh token, not an access token, for mobile refresh and logout.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: accessToken
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refreshToken
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
    widgetSiteKey:
      type: apiKey
      in: header
      name: X-Widget-Site-Key
  parameters:
...
    IntegrationKeyPath:
      name: integrationKey
      in: path
      required: true
      schema: { type: string, minLength: 1 }
...
    PageQuery: { name: page, in: query, required: false, schema: { type: integer, minimum: 1, default: 1 } }
    LimitQuery: { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
    CursorQuery: { name: cursor, in: query, required: false, schema: { type: integer } }
    ChannelQuery: { name: channel, in: query, required: false, schema: { type: string } }
    StatusQuery: { name: status, in: query, required: false, schema: { type: string } }
    DaysQuery: { name: days, in: query, required: false, schema: { type: integer } }
...
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: Must exactly match the canonical event's top-level eventId.
      schema: { type: string, minLength: 1 }
    WkilTimestampHeader:
      name: X-Wkil-Timestamp
      in: header
      required: true
      description: Unix timestamp in seconds used in the HMAC signing string.
      schema: { type: string, pattern: "^[0-9]+$", minLength: 1 }
    WkilSignatureHeader:
      name: X-Wkil-Signature
      in: header
      required: true
      description: HMAC-SHA256 signature in the form v1=<64 hexadecimal characters>; the current or previous signing secret may validate it during rotation overlap.
      schema: { type: string, pattern: "^v1=[0-9a-fA-F]{64}$" }
  responses:
    Ok:
      description: Successful response.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/GenericObject" }
    Created:
      description: Created response.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/GenericObject" }
...
    BadRequest:
      description: Bad request or validation failure.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Authentication required or failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Permission denied.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ServerError:
      description: Server error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    OrderEventAccepted:
      description: Event persisted and accepted for asynchronous processing. Duplicate events return the same shape with duplicate set to true.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/OrderEventAcceptedResponse" }
    OrderIntegrationList:
      description: Accessible order integrations.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/OrderIntegrationListResponse" }
    OrderIntegrationCreate:
      description: Created integration and one-time generated secret values.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/OrderIntegrationCreateResponse" }
...
  requestBodies:
...
    OrderIntegrationCreate:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/OrderIntegrationCreateRequest" }
...
    Lead:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/LeadRequest" }
...
    BusinessProfile:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/BusinessProfileRequest" }
...
    MediaUpload:
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            required: [businessProfileId, name, instructions, file]
            properties:
              businessProfileId: { type: integer }
              name: { type: string, minLength: 2 }
              instructions: { type: string, minLength: 10 }
              usageScope: { $ref: "#/components/schemas/MediaUsageScope" }
              file: { type: string, format: binary }
...
  schemas:
...
    GenericObject:
      type: object
      additionalProperties: true
    ErrorResponse:
      type: object
      additionalProperties: true
      properties:
        status: { type: string }
        message: { type: string }
        code: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              path: { type: string }
              message: { type: string }
...
    LeadRequest:
      type: object
      required: [name, email]
      properties:
        name: { type: string, minLength: 2, maxLength: 100 }
        email: { type: string, format: email }
        url: { type: string, format: uri }
        message: { type: string, maxLength: 1000 }
...
    BusinessProfileRequest:
      type: object
      required: [name, voice, tone, expectedUserIntents]
      properties:
        name: { type: string, minLength: 1 }
        voice: { type: string, minLength: 1 }
        tone: { type: string, minLength: 1 }
        expectedUserIntents: { type: array, minItems: 1, items: { type: string } }
        corePolicies: { type: string }
        documents:
          type: array
          default: []
          description: Initial knowledge documents (create only; updates are settings-only).
          items:
            type: object
            required: [kind, content]
            properties:
              kind: { type: string, pattern: "^[a-z][a-z0-9-]{0,31}$" }
              title: { type: string, maxLength: 200 }
              content: { type: string, minLength: 1, maxLength: 20000 }
        brandLogoUrl: { type: string, format: uri }
        brandPrimaryColor: { type: string, pattern: "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$" }
        brandSecondaryColor: { type: string, pattern: "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$" }
        brandAccentColor: { type: string, pattern: "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$" }
        visualAesthetic: { type: string }
        artStyle: { type: string }
        brandKitCompleted: { type: boolean, default: false }
        customerDetailsInstructions: { type: string }
        customerMemoryFields:
          type: array
          minItems: 3
          maxItems: 3
          items:
            type: object
            properties:
              key: { type: string }
              label: { type: string }
              description: { type: string }
        aiBehaviorInstructions: { type: string, maxLength: 4000 }
        handoffEnabled: { type: boolean, default: true }
        followUpEnabled: { type: boolean, default: false }
        followUpMode: { type: string, enum: [AUTO, CUSTOM], default: AUTO }
        followUpDelays: { type: array, maxItems: 5, items: { $ref: "#/components/schemas/FollowUpDelay" }, default: [] }
        followUpInstructions: { type: string, maxLength: 2000 }
...
    MediaUsageScope:
      type: string
      enum: [CHAT_ATTACHMENT, CONTENT_ASSET]
    OrderLocale:
      type: string
      enum: [ar, en]
...
    CanonicalOrderEvent:
      type: object
      additionalProperties: false
      required: [schemaVersion, eventId, eventType, occurredAt, order]
      properties:
        schemaVersion: { type: string, enum: ["1"] }
        eventId: { type: string, minLength: 1 }
        eventType: { type: string, enum: [order.created] }
        occurredAt: { type: string, format: date-time }
        order: { $ref: "#/components/schemas/CanonicalOrder" }
    CanonicalOrder:
      type: object
      additionalProperties: false
      required: [id, number, currency, total, customer]
      properties:
        id: { type: string, minLength: 1 }
        number: { type: string, minLength: 1 }
        currency: { type: string, pattern: "^[A-Z]{3}$" }
        total: { type: string, pattern: "^[0-9]+(?:\\.[0-9]+)?$" }
        customer: { $ref: "#/components/schemas/CanonicalOrderCustomer" }
        items:
          type: array
          items: { $ref: "#/components/schemas/CanonicalOrderItem" }
        shippingAddress: { $ref: "#/components/schemas/CanonicalShippingAddress" }
        sourceStatus: { type: string, minLength: 1 }
        paymentMethod: { type: string, minLength: 1 }
        metadata: { type: object, additionalProperties: true }
    CanonicalOrderCustomer:
      type: object
      additionalProperties: false
      required: [phone]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string, pattern: "^\\+[1-9][0-9]{1,14}$" }
        locale: { $ref: "#/components/schemas/OrderLocale" }
    CanonicalOrderItem:
      type: object
      additionalProperties: false
      required: [id, name, quantity, unitPrice, total]
      properties:
        id: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        quantity: { type: string, pattern: "^[0-9]+(?:\\.[0-9]+)?$" }
        unitPrice: { type: string, pattern: "^[0-9]+(?:\\.[0-9]+)?$" }
        total: { type: string, pattern: "^[0-9]+(?:\\.[0-9]+)?$" }
...
    OrderIntegrationCreateRequest:
      type: object
      additionalProperties: false
      required: [businessProfileId]
      properties:
        businessProfileId: { type: integer, minimum: 1 }
        whatsappAccountId: { type: [integer, "null"], minimum: 1 }
        defaultLocale: { $ref: "#/components/schemas/OrderLocale", default: en }
        isActive: { type: boolean, default: false }
        storeSyncEnabled: { type: boolean, default: false }
        statusCallbackUrl:
          type: [string, "null"]
          format: uri
          pattern: "^https://"
        statusCallbackSecret:
          type: [string, "null"]
          minLength: 1
          writeOnly: true
        callbackSecret:
          type: [string, "null"]
          minLength: 1
          writeOnly: true
          description: Alias for statusCallbackSecret.
        rotateStatusCallbackSecret: { type: boolean, default: false }
        rotateCallbackSecret:
          type: boolean
          default: false
          description: Alias for rotateStatusCallbackSecret.
...
    OrderEventAcceptedResponse:
      type: object
      additionalProperties: false
      required: [accepted, duplicate, eventId]
      properties:
        accepted: { type: boolean, const: true }
        duplicate: { type: boolean }
        eventId: { type: string, minLength: 1 }
    OrderIntegrationListResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/OrderIntegration" }
    OrderIntegrationCreateResponse:
      type: object
      additionalProperties: false
      required: [data, secret]
      properties:
        data: { $ref: "#/components/schemas/OrderIntegration" }
        secret:
          type: string
          minLength: 1
          readOnly: true
          description: Plaintext signing secret returned once at creation; never included in OrderIntegration.
        statusCallbackSecret:
          type: string
          minLength: 1
          readOnly: true
          description: Plaintext callback secret returned once only when callback-secret rotation was requested.
...